Improved Counter App: Individual Session Tracking

+++ counter/session.py
@@ -7,10 +7,12 @@
 import json
 import os
 import time
+import uuid
 
 import numpy as np
 
 from tools import pather
+from tools.logger import Logger
 
 __author__ = "Jakrin Juangbhanich"
 __copyright__ = "Copyright 2018, GenVis Pty Ltd."
@@ -23,49 +25,83 @@ class Session:
     OUTPUT_DIR = "output"
     Z_FILL_INDEX = 7
     ROLLING_WINDOW_SIZE = 10000  # This is how many session files we will keep.
-    SESSION_ACTIVATION_SEC = 0  # How many seconds before a session is counted as active.
     MAX_VECTOR_LENGTH = 10
+    SESSION_LONG_LIFE_FRAMES = 150  # How long to keep a session pending for, before ending it.
+    SESSION_SHORT_LIFE_FRAMES = 5  # How long to keep a session pending for (until it is full).
+    DISPLAY_TIME_LEFT_LIMIT = 0.9
+
+    VECTOR_COMPARE_LENGTH = 3
 
     def __init__(self):
         self.session_id = self.get_session_id()
-        self.face_id = "unknown"
+        self.face_id = uuid.uuid4().hex
         self.timestamp_start = time.time()
         self.timestamp_end = 0
+        self.local_time_start = time.localtime()
         self.vectors = []
+        self.time_left = self.SESSION_SHORT_LIFE_FRAMES
+        self.has_activated = False
+
+    @property
+    def display_id(self):
+        return "{}: {}".format(self.session_id, self.face_id[:6].upper())
 
     def add_vector(self, vector):
         self.vectors.append(vector)
         if len(self.vectors) > self.MAX_VECTOR_LENGTH:
             self.vectors.pop(0)
 
+        if not self.has_activated:
+            if self.is_full:
+                self.has_activated = True
+                Logger.field("Session Activated", "{}".format(self.display_id))
+
+        self.timestamp_end = time.time()
+        self.time_left = self.SESSION_LONG_LIFE_FRAMES if self.is_full else self.SESSION_SHORT_LIFE_FRAMES
+
+    @property
+    def is_full(self):
+        return len(self.vectors) == self.MAX_VECTOR_LENGTH
+
+    def update(self, time_delta):
+        self.time_left = max(0, self.time_left - time_delta)
+
+    @property
+    def time_left_percent(self):
+        return self.time_left/self.SESSION_LONG_LIFE_FRAMES
+
+    @property
+    def display_time_left_percent(self):
+        return min(1.0, self.time_left_percent / self.DISPLAY_TIME_LEFT_LIMIT)
+
     def get_distance(self, vector):
         distance_total = 0
-        for v in self.vectors:
+        cmp_length = min(len(self.vectors), self.VECTOR_COMPARE_LENGTH)
+        compare_vectors = self.vectors[:cmp_length]
+        for v in compare_vectors:
             distance_total += np.linalg.norm(v - vector)
         distance_total /= len(self.vectors)
         return distance_total
 
     def end(self):
         """ End the session and write the results to a file. """
-        self.timestamp_end = time.time()
+        Logger.field("Session Ended", "{}".format(self.display_id))
         self.create_results_data()
 
     @property
     def is_active(self):
-        return time.time() - self.timestamp_start >= self.SESSION_ACTIVATION_SEC
+        return self.time_left_percent > self.DISPLAY_TIME_LEFT_LIMIT
 
     def create_results_data(self):
         """ Create a dictionary of the results. """
-        if not self.is_active:
-            return
-
         data = {
             "session_id": self.session_id,
             "face_id": self.face_id,
-            "timestamp_start": self.timestamp_start,
-            "timestamp_end": self.timestamp_end,
-            "duration_in_seconds": self.timestamp_end - self.timestamp_start,
+            "timestamp_start": int(self.timestamp_start),
+            "timestamp_end": int(self.timestamp_end),
+            "duration_in_seconds": round(self.timestamp_end - self.timestamp_start, 2),
             "date": self._get_readable_date(time.localtime()),
+            "readable_time_start": self._get_readable_time(self.local_time_start),
             "readable_time_end": self._get_readable_time(time.localtime())
         }
 
@@ -91,10 +127,13 @@ class Session:
 
         # Delete the first n files from the session folder.
         excess = len(session_files) - rolling_window_size
-        prune_files = session_files[:excess]
-        for f in prune_files:
-            file_path = os.path.join(self.OUTPUT_DIR, f)
-            os.remove(file_path)
+        Logger.field("File Storage", "{}/{}".format(len(session_files), rolling_window_size))
+        if excess > 0:
+            prune_files = session_files[:excess]
+            for f in prune_files:
+                Logger.field("Pruning File", "{}".format(f))
+                file_path = os.path.join(self.OUTPUT_DIR, f)
+                os.remove(file_path)
 
     def _create_session_file(self):
         """ Create the session file if it doesn't yet exist. """
